fix: Update calls to unstructured-client for forward compatibility
Per the usage example [here](https://docs.unstructured.io/api-reference/api-services/sdk-python#basics), the call to `sdk.general.partition` expects a `shared.PartitionRequest` object. In older versions of the client, this object was not needed, but the call has stayed backwards compatible. The next version of the client (0.26.0) is going to lose support for directly passing `PartitionParameters`.