Properly log elasticsearch upload errors
Original error logging was never called because by default parallel_bulk
re-raises exceptions and raises errors for non 2XX responses and these
were not caught.

We change the logic to catch, log and re-raise errors on our side.
Error log is sanitized to remove the uploaded object contents from it.