fix/NEXUS-777: Fix breaking change in upcoming unstructured-client
In the unreleased 0.30.0 of `unstructured-client`, handling of the `server_url` is changing. We need
to do this to integrate the platform API functions. Now, different parts of the SDK are talking to
different backend services, and so it's preferred that we set the `server_url` per endpoint.

Note that for compatibility with the current SDK, we need to strip the `/general/v0/general` path
off of the URL. We have logic to do this in the SDK, but in versions before 0.30.0 this doesn't take
effect when you pass the URL in the endpoint like this.