[P6M-1327]: Make user_pname optional for Sharepoint (#507)
Due to Sharepoint inheriting Onedrive, we originally had user_pname
required. But it isn't really required.

It's actually not even used unless the User user username/password as
their form of Auth which is not recommended by Microsoft.

SharepointConnectionConfig inherits from OnedriveConnectionConfig ,
which is where the use and declaration of this arg is made. With the
changes in the PR, we override the arg to be Optional


Also updated the vertexai model which was deprecated.