fix: change default to pinecone serverless_region default (#522)
> I'm using the "opinionated writes" feature for the Pinecone
destination connector and getting this error: Your free plan does not
support indexes in the us-west-2 region of aws. To create indexes in
this region, upgrade your plan. But I don't want to upgrade my free
plan. Is there any way to force this index to be created in us-east-1
instead?

Per [the docs on Regions available for serverless
indexes](https://docs.pinecone.io/troubleshooting/available-cloud-regions#regions-available-for-serverless-indexes),
we need to change our default to a starter plan friendly value. This PR
addresses that