Allow setting REQUESTS_CA_BUNDLE to use custom certs for OpenAI embedders (#561)
If your embedding models are hosted within an enterprise/private PKI
environment, you many need to trust a custom certificate bundle when
interacting with them. This change allows setting REQUESTS_CA_BUNDLE to
a file which will be used by the HTTP client in OpenAI and Azure OpenAI
embedding interactions.

---------

Co-authored-by: ryannikolaidis <1208590+ryannikolaidis@users.noreply.github.com>