remove deletion operation in es connector (#570)
in ES connector, function delete_by_record_id is not necessary, because
when document_id already exists in the index, we want to replace the
record with new content, which is equivalent to upsert operation.
deleting existing records is dangerous because it may remove unintended
content inadvertently. For more details see:

https://linear.app/unstructured/issue/ENG-360/upsert-in-elastic-incorrectly-implemented