feat: Add binary_encode_vectors flag to AstraDBUploader
AstraDB uploads vectors with their own binary format by default. This is more efficient, but it
makes it hard to view and work with vectors directly in the UI. We can add a flag to turn off this
encoding if you're willing to take the performance hit.

Tried adding an integration test, but vectors are also read in binary format so this doesn't verify
anything. Instead, let's start a unit test suite for Astra to confirm that the option is being set.